Refactor GraphQL resolvers and remove unused server file
