+++ app/graphql/schema/queries/get-full-address.js
@@ -1,0 +1,7 @@
+const getFullAddress = (address) => {
+  return Object.values(address).filter(x => x).join(', ')
+}
+
+module.exports = {
+  getFullAddress
+}
