+++ app/graphql/schema/queries/index.js
@@ -1,0 +1,11 @@
+const { personOrganisations } = require('./person-organisations')
+const { person } = require('./person')
+const { organisation } = require('./organisation')
+const { permissions } = require('./permissions')
+
+module.exports = {
+  personOrganisations,
+  person,
+  organisation,
+  permissions
+}
