+++ app/graphql/schema/queries/person-organisations.js
@@ -1,0 +1,17 @@
+const { get } = require('../../../api')
+
+const personOrganisations = async (_root, _args, context) => {
+  const response = await get('/organisation/person/3337243/summary?search=', context.crn, context.token)
+  return {
+    crn: context.crn,
+    organisations: response._data?.map(x => ({
+      id: x.id,
+      sbi: x.sbi,
+      name: x.name
+    })) ?? []
+  }
+}
+
+module.exports = {
+  personOrganisations
+}
