+++ package.json
@@ -14,12 +14,17 @@
     "start:debug": "nodemon --inspect-brk=0.0.0.0 --ext js --legacy-watch app/index.js"
   },
   "author": "Defra",
-  "contributors": [],
+  "contributors": [
+    "John Watson john.watson1@defra.gov.uk",
+    "Fay Toward fay.toward@defra.gov.uk",
+    "Simon Dunn simon.dunn1@defra.gov.uk",
+    "Rana Salem rana.salem@defra.gov.uk"
+  ],
   "license": "OGL-UK-3.0",
   "dependencies": {
     "@as-integrations/hapi": "2.0.1",
     "@hapi/boom": "10.0.1",
-    "@hapi/catbox-redis": "^6.0.2",
+    "@hapi/catbox-redis": "6.0.2",
     "@hapi/hapi": "21.3.2",
     "applicationinsights": "2.9.1",
     "joi": "17.12.0"
